# source/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Document test API Locofull",
        "description": "",
        "termsOfService": "",
        "contact": {
            "email": "tien.nguyen@saver.jp"
        },
        "license": {
            "name": "Saver",
            "url": "https://www.saver.jp/"
        },
        "version": "1.0.0"
    },
    "basePath": "/backend/api",
    "paths": {
        "/admin/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "email",
                        "required": true,
                        "type": "string",
                        "default": "lanh.nguyen1@saver.jp"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password",
                        "required": true,
                        "type": "string",
                        "default": "123456789"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{<pre>&emsp;  'success': true | false,\r\n    &emsp;  'data': {\r\n    &emsp;            'id': 1,\r\n    &emsp;            'api_token': 'eFAQOXsVkudj74VBaC0prLrbqBXP7U6TS3KwcJhQYb4SN7oxloWhgL0lMO2E',\r\n    &emsp;            'username': null,\r\n    &emsp;            'email': null\r\n    &emsp;          } | null,\r\n    &emsp;  'error': null | ...</pre>}\r\n    "
                    },
                    "404": {
                        "description": "404 page"
                    }
                }
            }
        },
        "/admin/logout": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Logout Admin",
                "parameters": [
                    {
                        "name": "admin_id",
                        "in": "formData",
                        "description": "Id of Admin",
                        "required": true,
                        "type": "number",
                        "default": "1"
                    },
                    {
                        "name": "api_token",
                        "in": "formData",
                        "description": "API token",
                        "required": true,
                        "type": "string",
                        "default": "eFAQOXsVkudj74VBaC0prLrbqBXP7U6TS3KwcJhQYb4SN7oxloWhgL0lMO2E"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{<pre>&emsp;  'success': true | false,\r\n    &emsp;  'data': null,\r\n    &emsp;  'error': null | ...</pre>}\r\n    "
                    },
                    "404": {
                        "description": "404 page"
                    }
                }
            }
        },
        "/admin/add": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Add",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "formData",
                        "description": "API token",
                        "required": true,
                        "type": "string",
                        "default": "lypqyENjHEhGZ2Grt45HNiOCyb9vunjzfZYtuxV3r2RZaPFc05X2B4MbEwpv"
                    },
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "Username",
                        "required": true,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password",
                        "required": true,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "password_confirmation",
                        "in": "formData",
                        "description": "Password Confirmation",
                        "required": true,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "phone_number",
                        "in": "formData",
                        "description": "Phone number",
                        "type": "number",
                        "default": ""
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email",
                        "type": "string",
                        "default": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{<pre>&emsp;  'success': true | false,\r\n    &emsp;  'data': 1 | null,\r\n    &emsp;  'error': null | ...</pre>}\r\n    "
                    },
                    "404": {
                        "description": "Building not found"
                    }
                }
            }
        },
...
        "/admin/edit/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Edit",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "formData",
                        "description": "API token",
                        "required": true,
                        "type": "string",
                        "default": "lypqyENjHEhGZ2Grt45HNiOCyb9vunjzfZYtuxV3r2RZaPFc05X2B4MbEwpv"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "type": "string",
                        "default": "3"
                    },
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "Username",
                        "required": true,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password",
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "password_confirmation",
                        "in": "formData",
                        "description": "Password Confirmation",
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "phone_number",
                        "in": "formData",
                        "description": "Phone number",
                        "type": "number",
                        "default": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{<pre>&emsp;  'success': true | false,\r\n    &emsp;  'data': '5',\r\n    &emsp;  'error': null | ...</pre>}\r\n    "
                    },
                    "404": {
                        "description": "Building not found"
                    }
                }
            }
        },
...
    },
...
}